requirements.txt
anyio==4.1.0
certifi==2023.11.17
dnspython==2.4.2
exceptiongroup==1.2.0
h11==0.14.0
httpcore==1.0.2
httpx==0.25.2
idna==3.6
pymongo==4.6.1
python-telegram-bot==20.7
sniffio==1.3.0
